fix: address PR review findings (edge cases + potential bugs)
Confirmed via failing-first tests, then fixed:
- CounterAggregator summed Counter.Value in checked Int32; a hot adapter's
  un-collapsed :dur key (per-call latency ms) could exceed int.MaxValue and
  throw OverflowException, wedging the whole counter→statistic pipeline (jobs
  included). Sum as long, matching the read path. (test: PG+SqlServer)
- Observe-first adapters recorded null StatusCode: SetStatusCode sat behind
  the capture-mode guard, but status code is always-recorded metadata (§8.19).
  Set it before the guard so a no-capture adapter can still distinguish 404
  from 500. (test: NoDb; existing test updated)
- Webhook group drill-down double-decoded its route param (React Router 7
  already decodes useParams), corrupting/crashing on endpoint-URL keys with
  '%'. Use the param as-is.

Hardening:
- ExpirationCleanup count sweeps get a ThenBy(Id) tiebreaker so the keep-set
  among equal timestamps is deterministic.
- AdapterRateLimiter.FloorWindow guards a non-positive window (admin override
  WindowSeconds<=0) instead of dividing by zero.
- Inbound endpoint recorder now counts channel-full drops
  (warp.endpoint.records_dropped), mirroring the adapter recorder.

Investigated and NOT a bug (regression test added): webhook delivery-history
hour bucketing — Npgsql translates the timestamptz DateTime-part GroupBy to
date_part(.. AT TIME ZONE 'UTC'), so buckets are UTC-correct regardless of the
PG session TimeZone; the raw-date_part concern doesn't reach EF.

All green: Adapters/Endpoints/Webhooks — NoDb 151, PG 194, SqlServer 193;
Warp.Core + UI analyzer/type-clean.

+++ src/core/Warp.Worker/Services/CounterAggregator.cs
@@ -49,7 +49,7 @@ public sealed class CounterAggregator<TContext> : IServerTask
 
         var grouped = counters
             .GroupBy(x => x.Key)
-            .Select(g => new { Key = g.Key, Sum = g.Sum(x => x.Value) });
+            .Select(g => new { Key = g.Key, Sum = g.Sum(x => (long)x.Value) });
 
         foreach (var group in grouped)
         {
